+++ src/main/java/com/offlineplayersreworked/exception/InvalidOffsetException.java
@@ -1,0 +1,7 @@
+package com.offlineplayersreworked.exception;
+
+public class InvalidOffsetException extends IllegalArgumentException {
+    public InvalidOffsetException(String s) {
+        super(s);
+    }
+}