+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-# OfflinePlayers-Reworked 1.1.2 (MC 1.21.1)
+# OfflinePlayers-Reworked 1.2.0 (MC 1.21.1)
 
 **Update:**
 
-- Fixed bug where sun gets stuck at sunrise when a player is in offline mode
+- Changed package to com.offlineplayersreworked
+- Added Gametest
+- Updated to loader version 0.18.4
+- Updated gradle to 9.2.1